HTKJC-ST01-011: split tests
